requirements.txt
# 核心Web框架
fastapi>=0.115.0,<0.116.0
uvicorn>=0.24.0
websockets>=12.0
starlette>=0.46.0,<0.47.0

# 数据验证和配置
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
typing-extensions>=4.8.0

# LangChain和AI相关
langchain-core==0.3.72
langchain-openai==0.3.28
langchain-tavily==0.2.11
langchain-text-splitters==0.3.9
langgraph==0.6.2
langgraph-checkpoint==2.1.1
langgraph-checkpoint-postgres==2.0.21
langgraph-prebuilt==0.6.2
langgraph-sdk==0.2.0
langmem==0.0.29
langsmith==0.4.8
ffmpeg==1.4



# 数据库相关
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.9
psycopg-binary>=3.2.0
psycopg-pool>=3.2.0

# 认证和安全
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
PyJWT>=1.7.1,<2.10.0
jwt>=1.3.1

# 邮件和验证
email-validator>=2.0.0
sendgrid>=6.10.0

# HTTP客户端和请求
aiohttp>=3.8.0
requests>=2.31.0
httpx>=0.27.0,<0.28.0
python-multipart>=0.0.6

# 配置和环境
python-decouple>=3.8

# 任务队列
redis>=5.0.0
celery>=5.3.0

# 图像处理
Pillow>=10.0.0
opencv-python>=4.8.0
